fix #119897

+++ cache2k-core/src/main/java/org/cache2k/core/ClockProPlusEviction.java
@@ -246,7 +246,7 @@ public class ClockProPlusEviction extends AbstractEviction {
       hand = hand.next;
     }
     this.hotHits = hotHits;
-    long scanCount = initialMaxScan - maxScan;
+    long scanCount = (long)initialMaxScan - (long)maxScan;
     hotScanCnt += scanCount;
     handHot = hand;
     return coldCandidate;
